Backend/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat @ file:///tmp/build/80754af9/automat_1600298431173/work
bcrypt @ file:///C:/Windows/Temp/abs_36kl66t_aw/croots/recipe/bcrypt_1659554334050/work
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.5
daphne==4.0.0
Django==3.2.23
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djongo==1.3.6
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
hyperlink @ file:///tmp/build/80754af9/hyperlink_1610130746837/work
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
incremental @ file:///tmp/build/80754af9/incremental_1636629750599/work
install==1.3.5
jwt==1.3.1
msgpack==1.0.7
mssql-django==1.3
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyHamcrest @ file:///tmp/build/80754af9/pyhamcrest_1615748656804/work
PyJWT==2.8.0
pymongo==3.12.0
pyodbc @ file:///C:/Windows/Temp/abs_61e3jz3u05/croots/recipe/pyodbc_1659513801402/work
pyOpenSSL==23.3.0
pytz==2021.3
pywin32==305.1
redis==5.0.1
service-identity @ file:///Users/ktietz/demo/mc3/conda-bld/service_identity_1629460757137/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse==0.2.4
Twisted @ file:///C:/Windows/Temp/abs_ccblv2rzfa/croots/recipe/twisted_1659592764512/work
twisted-iocpsupport @ file:///C:/ci/twisted-iocpsupport_1646817106095/work
txaio==23.1.1
typing_extensions==4.7.1
wincertstore==0.2
zope.interface @ file:///C:/ci/zope.interface_1625036257680/work
